requirements.txt
streamlit
openai
python-dotenv
pandas
pillow
pypdf
openpyxl
gtts
streamlit-webrtc
duckduckgo-search
soundfile
numpy
python-docx